fix: moves around a bunch of files to separate ui-shell from Odyssey
