+++ packages/odyssey-react-mui/src/index.ts
@@ -71,7 +71,6 @@ export * from "./Checkbox";
 export * from "./CheckboxGroup";
 export * from "./CircularProgress";
 export * from "./ContrastModeProvider";
-export * from "./createMessageBus";
 export * from "./createUniqueId";
 export * from "./CssBaseline";
 export * from "./DataTable";
@@ -101,12 +100,9 @@ export * from "./Pagination";
 export * from "./PasswordField";
 export * from "./Radio";
 export * from "./RadioGroup";
-export * from "./renderOktaUiShell";
-export * from "./renderReactInWebComponent";
 export * from "./ScreenReaderText";
 export * from "./SearchField";
 export * from "./Select";
-export * from "./shadow-dom";
 export * from "./Stack";
 export * from "./Status";
 export * from "./Surface";
@@ -121,3 +117,4 @@ export * from "./ToastStack";
 export * from "./Tooltip";
 export * from "./Typography";
 export * from "./useUniqueId";
+export * from "./web-component";
