+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -37,3 +37,4 @@ export * from "./GroupPicker";
 export * from "./NavAccordion";
 export * from "./SideNav";
 export * from "./TopNav";
+export * from "../ui-shell";
