+++ packages/odyssey-react-mui/src/test-utils/captureConsoleError.ts
