+++ packages/odyssey-react-mui/src/web-component/index.ts
@@ -10,4 +10,5 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-export { renderOktaUiShell } from "@okta/odyssey-react-mui/labs";
+export * from "./renderReactInWebComponent";
+export * from "./shadow-dom";
